refactor(frontend): split BillCalendarPage into bill-calendar/ module
